ДЗ #3: Rest-assured -  исправление замечаний.
