--- src/main/java/ru/autotest/Main.java
@@ -1,7 +1,0 @@
-package ru.autotest;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
