+++ src/main/java/ru/autotest/data/StatusPetEnum.java
@@ -1,7 +1,7 @@
 package ru.autotest.data;
 
 public enum StatusPetEnum {
-    available, pending, sold
-
-
+    AVAILABLE,
+    PENDING,
+    SOLD
 }
