+++ src/main/java/ru/autotest/dto/ResponseErrorDTO.java
@@ -1,14 +1,10 @@
 
 package ru.autotest.dto;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.google.gson.annotations.SerializedName;
-import lombok.Builder;
 import lombok.Data;
 
 @Data
-@SuppressWarnings("unused")
-//@JsonIgnoreProperties(ignoreUnknown = true)
 public class ResponseErrorDTO {
 
     @SerializedName("code")
