--- src/test/java/ru/autotest/MainTest.java
@@ -1,4 +1,0 @@
-package ru.autotest;
-
-public class MainTest {
-}
